Update wording on function to padd new inputs to mode

+++ src/mode.h
@@ -33,11 +33,11 @@ class Mode {
         virtual void initialise_pot_4(unsigned long current_millis) {}
         virtual void initialise_button_0(unsigned long current_millis) {}
 
-        virtual void process_new_pot_0_value(uint16_t value) {}
-        virtual void process_new_pot_1_value(uint16_t value) {}
-        virtual void process_new_pot_2_value(uint16_t value) {}
-        virtual void process_new_pot_3_value(uint16_t value) {}
-        virtual void process_new_pot_4_value(uint16_t value) {}
+        virtual void process_new_pot_0_value(int value) {}
+        virtual void process_new_pot_1_value(int value) {}
+        virtual void process_new_pot_2_value(int value) {}
+        virtual void process_new_pot_3_value(int value) {}
+        virtual void process_new_pot_4_value(int value) {}
         virtual void button_0_pressed() {}
         virtual void button_0_released() {}
     protected:
